bump to v0.0.5

+++ migrations/20240303072633_create_twitter_token.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS twitter_tokens (
     id INT PRIMARY KEY NOT NULL,
     token_type VARCHAR,
-    expires_in INT,
+    expires_in INT8,
     access_token VARCHAR NOT NULL,
     scope VARCHAR,
     refresh_token VARCHAR NOT NULL
